// src/generated/drift/types/spotPosition.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  fixDecoderSize,
  fixEncoderSize,
  getBytesDecoder,
  getBytesEncoder,
  getI64Decoder,
  getI64Encoder,
  getStructDecoder,
  getStructEncoder,
  getU16Decoder,
  getU16Encoder,
  getU64Decoder,
  getU64Encoder,
  getU8Decoder,
  getU8Encoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
  type ReadonlyUint8Array,
} from "@solana/kit";
import {
  getSpotBalanceTypeDecoder,
  getSpotBalanceTypeEncoder,
  type SpotBalanceType,
  type SpotBalanceTypeArgs,
} from ".";

export type SpotPosition = {
  /**
   * The scaled balance of the position. To get the token amount, multiply by the cumulative deposit/borrow
   * interest of corresponding market.
   * precision: SPOT_BALANCE_PRECISION
   */
  scaledBalance: bigint;
  /**
   * How many spot non reduce only trigger orders the user has open
   * precision: token mint precision
   */
  openBids: bigint;
  /**
   * How many spot non reduce only trigger orders the user has open
   * precision: token mint precision
   */
  openAsks: bigint;
  /**
   * The cumulative deposits/borrows a user has made into a market
   * precision: token mint precision
   */
  cumulativeDeposits: bigint;
  /** The market index of the corresponding spot market */
  marketIndex: number;
  /** Whether the position is deposit or borrow */
  balanceType: SpotBalanceType;
  /** Number of open orders */
  openOrders: number;
  padding: ReadonlyUint8Array;
};

export type SpotPositionArgs = {
  /**
   * The scaled balance of the position. To get the token amount, multiply by the cumulative deposit/borrow
   * interest of corresponding market.
   * precision: SPOT_BALANCE_PRECISION
   */
  scaledBalance: number | bigint;
  /**
   * How many spot non reduce only trigger orders the user has open
   * precision: token mint precision
   */
  openBids: number | bigint;
  /**
   * How many spot non reduce only trigger orders the user has open
   * precision: token mint precision
   */
  openAsks: number | bigint;
  /**
   * The cumulative deposits/borrows a user has made into a market
   * precision: token mint precision
   */
  cumulativeDeposits: number | bigint;
  /** The market index of the corresponding spot market */
  marketIndex: number;
  /** Whether the position is deposit or borrow */
  balanceType: SpotBalanceTypeArgs;
  /** Number of open orders */
  openOrders: number;
  padding: ReadonlyUint8Array;
};

export function getSpotPositionEncoder(): FixedSizeEncoder<SpotPositionArgs> {
  return getStructEncoder([
    ["scaledBalance", getU64Encoder()],
    ["openBids", getI64Encoder()],
    ["openAsks", getI64Encoder()],
    ["cumulativeDeposits", getI64Encoder()],
    ["marketIndex", getU16Encoder()],
    ["balanceType", getSpotBalanceTypeEncoder()],
    ["openOrders", getU8Encoder()],
    ["padding", fixEncoderSize(getBytesEncoder(), 4)],
  ]);
}

export function getSpotPositionDecoder(): FixedSizeDecoder<SpotPosition> {
  return getStructDecoder([
    ["scaledBalance", getU64Decoder()],
    ["openBids", getI64Decoder()],
    ["openAsks", getI64Decoder()],
    ["cumulativeDeposits", getI64Decoder()],
    ["marketIndex", getU16Decoder()],
    ["balanceType", getSpotBalanceTypeDecoder()],
    ["openOrders", getU8Decoder()],
    ["padding", fixDecoderSize(getBytesDecoder(), 4)],
  ]);
}

export function getSpotPositionCodec(): FixedSizeCodec<SpotPositionArgs, SpotPosition> {
  return combineCodec(getSpotPositionEncoder(), getSpotPositionDecoder());
}